+++ NEWS.md
@@ -10,6 +10,10 @@ News for Package 'SDLfilter'
 
 #### Version 2.1.1 (2021-07-04)
 
+* added an option to estimate the confidence intervals for bootstrapped overlap estimates. The new default is to use the CIs when considering an asymptote.
+
+#### Version 2.1.1 (2021-07-04)
+
 * fixed a minor bug related with the max.asymptote argument within the asymptote function.
 
 #### Version 2.1.1 (2021-07-04)
@@ -21,7 +25,7 @@ News for Package 'SDLfilter'
 * added a new citation for functions 'boot_overlap', 'boot_area' and 'asymptote'. Shimada T, Thums M, Hamann M, Limpus CJ, Hays GC, FitzSimmons NN, Wildermann NE, Duarte CD, Meekan MG (2021) Optimising sample sizes for animal distribution analysis using tracking data. Methods in Ecology and Evolution, 12(2):288-297. doi:10.1111/2041-210X.13506
 * updated the README file.
 * improved the R documentations of functions asymptote, boot_overlap, and boot_area.
-* remaned the example data "curtis" to "flatback".
+* renamed the example data "curtis" to "flatback".
 
 
 #### Version 2.0.1 (2020-07-04)